refactor energyBalanceGraph
