+++ app/frontend/components/Graph/EnergyBalanceChart.tsx
@@ -1,0 +1,78 @@
+import {
+    Bar,
+    BarChart,
+    CartesianGrid,
+    Label,
+    ResponsiveContainer,
+    Tooltip,
+    XAxis,
+    YAxis,
+} from 'recharts';
+import { EnergyBalanceChartProps } from '@/types/components/Graph';
+
+
+export default function EnergyBalanceChart({
+    graphData,
+    legendData,
+    metaData,
+}: EnergyBalanceChartProps) {
+    const getUniqueKeys = () => {
+        const uniqueKeys = [];
+        graphData.map((item) => {
+            Object.keys(item).map((key) => {
+                if (!uniqueKeys.includes(key)) {
+                    uniqueKeys.push(key);
+                }
+            });
+        });
+        return uniqueKeys;
+    };
+
+    if (graphData.length === 0) {
+        return (
+            <div className="flex items-center justify-center h-full text-gray-500">
+                Geen data beschikbaar
+            </div>
+        );
+    }
+
+    return (
+        <ResponsiveContainer width="100%" height="100%">
+            <BarChart
+                width={500}
+                height={300}
+                data={graphData}
+                margin={{
+                    top: 20,
+                    right: 30,
+                    left: 20,
+                    bottom: 5,
+                }}>
+                <CartesianGrid strokeDasharray="3 3" />
+                <XAxis dataKey="name" />
+                <YAxis>
+                    <Label angle={-90} dx={-20}>
+                        {`${metaData.title} (${metaData.unit})`}
+                    </Label>
+                </YAxis>
+                <Tooltip
+                    formatter={(value: number, name: string) => [
+                        value.toFixed(2),
+                        name,
+                    ]}
+                />
+                {getUniqueKeys()
+                    .filter((bar) => bar !== 'name')
+                    .map((item, index) => (
+                        <Bar
+                            key={index}
+                            isAnimationActive={false}
+                            dataKey={item}
+                            stackId="a"
+                            fill={legendData[item]}
+                        />
+                    ))}
+            </BarChart>
+        </ResponsiveContainer>
+    );
+}
